refactor (naming): drop the app_ prefix from files and classes
AppIcon becomes SvgIcon (a literal Icon would collide with Flutter's;
the new name says what the widget is — the one door SVGs enter
through), AppStartupWidget becomes StartupWidget in widgets/startup/,
the appStartup provider becomes startup in startup_provider.dart, and
the localization extension file and name lose the prefix. Exempt:
app_bar_theme.dart (about AppBar, not a prefix) and the generated
AppLocalizations (renaming collides with Flutter's Localizations).
