+++ lib/src/core/utiliity/validation/required_validation.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../extensions/app_localization.dart';
+import '../../extensions/localization.dart';
 import 'validation.dart';
 
 class RequiredValidation<T> extends Validation<T> {
