+++ lib/src/presentation/core/application_state/startup_provider/startup_provider.dart
@@ -3,10 +3,10 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 import '../../../../core/di/dependency_injection.dart';
 import '../localization_provider/localization_provider.dart';
 
-part 'app_startup_provider.g.dart';
+part 'startup_provider.g.dart';
 
 @Riverpod(keepAlive: true)
-Future<void> appStartup(Ref ref) async {
+Future<void> startup(Ref ref) async {
   ref.onDispose(() {
     ref.invalidate(sharedPreferencesProvider);
   });
