+++ lib/src/presentation/core/widgets/not_found_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../core/extensions/app_localization.dart';
+import '../../../core/extensions/localization.dart';
 import '../router/routes.dart';
 import '../theme/theme.dart';
