+++ lib/src/presentation/core/widgets/startup/startup_widget.dart
@@ -1,29 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../application_state/startup_provider/app_startup_provider.dart';
+import '../../application_state/startup_provider/startup_provider.dart';
 import '../failure_view.dart';
 
-class AppStartupWidget extends ConsumerWidget {
-  const AppStartupWidget({
-    super.key,
-    required this.loading,
-    required this.loaded,
-  });
+class StartupWidget extends ConsumerWidget {
+  const StartupWidget({super.key, required this.loading, required this.loaded});
 
   final Widget loading;
   final Widget loaded;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final appStartupState = ref.watch(appStartupProvider);
+    final startupState = ref.watch(startupProvider);
 
-    return appStartupState.when(
+    return startupState.when(
       loading: () => loading,
       error: (error, stackTrace) => Scaffold(
         body: FailureView(
           error: error,
-          onRetry: () => ref.invalidate(appStartupProvider),
+          onRetry: () => ref.invalidate(startupProvider),
         ),
       ),
       data: (_) => loaded,
