+++ lib/src/presentation/features/authentication/forgot_password/view/create_new_password_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../core/extensions/app_localization.dart';
+import '../../../../../core/extensions/localization.dart';
 import '../../../../core/router/routes.dart';
 import '../../../../core/theme/theme.dart';
 import '../../../../core/widgets/text/typography.dart';
