+++ lib/src/presentation/features/authentication/login/view/login_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../core/extensions/app_localization.dart';
+import '../../../../../core/extensions/localization.dart';
 import '../../../../../core/extensions/validation.dart';
 import '../../../../../core/utiliity/validation/validation.dart';
 import '../../../../../domain/failures/business_failure.dart';
