+++ lib/src/presentation/features/authentication/login/widgets/language_switcher.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
-import '../../../../../core/extensions/app_localization.dart';
+import '../../../../../core/extensions/localization.dart';
 import '../../../../../core/gen/l10n/app_localizations.dart';
 import '../../../../core/application_state/localization_provider/localization_provider.dart';
 import '../../../../core/theme/theme.dart';
